add: implementacao-telaRegistro1
