--- src/components/ConteudoDireito1/conteudoDireito.css
@@ -1,17 +1,0 @@
-*{
-    margin: 0;
-    padding: 0;
-}
-
-.header{
-    display: flex;
-    
-}
-
-
-.entrar{
-    color: #006845;
-    text-decoration:none; 
-    font-weight: 500;
-
-}
