+++ src/components/telaRegistro1/telaRegistro1.css
@@ -1,0 +1,142 @@
+*{
+    margin: 0;
+    padding: 0;
+}
+
+.ConteudoDireito{
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    padding: 20px;
+    width: 100%;
+    height: 100%;
+}
+
+.header{
+    display: flex;
+    justify-content: end;
+    width: 100%;
+}
+
+.entrar{
+    color: #006845;
+    text-decoration: none;
+    font-weight: 500;
+}
+
+.container-conteudoDireito{
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    height: 100%;
+    justify-content: center;
+    gap: 10px;
+}
+
+/* Estilização do botão  */
+
+.botaoCompleto{
+    box-shadow: rgba(0, 0, 0, 0.15) 1.95px 1.95px 2.6px;
+    display: flex;
+    align-items: center;
+    border-radius: 25px;
+    background-color: #FFFFFF;
+    height: 100px;
+    width: 90%;
+    border: 1px solid #006845;
+    cursor: pointer;
+    padding: 15px;
+    transition: 1s;
+}
+
+.botaoCompleto:hover{
+    background-color: #006845;
+    color: white;
+    box-shadow: rgba(3, 3, 3, 0.80) 1.95px 1.95px 2.6px;
+}
+
+.descricaoTexto{
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    width: 100%;
+    text-align: start;  
+}
+
+.descricaoTexto h4{
+    font-size: 20px;
+    font-style: normal;
+    font-weight: 500;
+    line-height: normal;
+    margin: 0px;
+}
+
+.descricaoTexto p{
+    color: #8692A6;
+    font-size: 14px;
+    font-weight: 400;
+    margin: 0px;
+    transition: 1s;
+}
+
+.descricaoTexto p:hover{
+    color: white;
+}
+
+.iconeTipoConta{
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    background-color: #008753;;
+    padding: 10px;
+    margin-right: 10px;
+    color: white;
+    clip-path: polygon(50% 0%, 100% 38%, 82% 100%, 18% 100%, 0% 38%);
+}
+
+.avanca{
+    padding-left: 20px;
+}
+
+@media screen and (max-width: 1300px) , (max-height: 700px){
+    
+.botaoCompleto{
+    height: 80px;
+}
+
+.descricaoTexto h4{
+    font-size: 18px;
+    transition: 1s;
+}
+
+.descricaoTexto p{
+    color: #8692A6;
+    font-size: 13px;
+    font-weight: 400;
+    margin: 0px;
+    transition: 1s;
+}
+
+}
+
+
+@media screen and (max-width: 1100px) , (max-height: 600px){
+    
+    .botaoCompleto{
+        height: 100px;
+    }
+    
+    .descricaoTexto h4{
+        font-size: 18px;
+        transition: 1s;
+    }
+    
+    .descricaoTexto p{
+        color: #8692A6;
+        font-size: 13px;
+        font-weight: 400;
+        margin: 0px;
+        transition: 1s;
+    }
+    
+}
