+++ src/components/telasLogin/LadoEsquerdoRegistro.css
@@ -1,6 +1,7 @@
+/* Testes com o framework: Swiper */
 .tentativadeSwiper{
-    width: 100vw;
-    height: 100vh;
+    width: 50%;
+    height: 100%;
     display: flex;
     justify-content: center;
     align-items: center;
@@ -10,8 +11,8 @@
     display: flex;
     justify-content: center;
     align-items: center;
-    width: 30vw;
-    height: 80vh;
+    width: 100%;
+    height: 100%;
 }
 
 .tamanhoImage{
